\name{vbCPP-package}
\alias{vbCPP-package}
\alias{vbCPP}
\docType{package}
\title{
\packageTitle{vbCPP}
}
\description{
\packageDescription{vbCPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vbCPP}
\packageIndices{vbCPP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{vbCPP}

Maintainer: \packageMaintainer{vbCPP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
